Made triggering available through mixins
...and changed the implementation to let the triggered module not be
aware of what to do, instead call a callback function on the triggering
module and let that do something with the audioParam.
This enables the QwertyHancock to be able to trigger an EG which in turn
triggers a Gain. It also prepares for letting EGs also trigger Osc and
Filter modules (need to extend to have the triggered module tell the
triggering module what values to go from and to)